new plugin update - ANdroid
