+++ pubspec.yaml
@@ -5,7 +5,7 @@ homepage: https://kommunicate.io
 
 environment:
   sdk: '>=2.12.0 <3.0.0'
-  flutter: ">=1.10.0"
+  flutter: "1.12.0"
 
 dependencies:
   flutter:
